tok64.c: Include stdlib.h, int main and constification
<stdlib.h> is required for atoi().

main() must return int not void , so we declare it that way. This
is fine because it does an explicit exit() at the end anyway.

The structures to which I added `const` are static structures of
keywords and the like which look like they should not modified at
runtime. Probably these were not originally declared `const` because
the code was written in in the '90s.

The strings to which I added `const` are clearly meant to be static.

disp_list() also changes its parameter declarations to use `const`
for the strings it does not modify.

All these fix errors emitted by GCC 6.3 on Debian 9.