GETLINE.H: Drop `\r\n` as well as `\n`, for POSIX compatibility
On POSIX systems (as tested on Linux), getline() was stripping off the
`\n` on native format files but leaving the `\r` at the end if the
file was in DOS format (CR-NL line terminators). This `\r` was then
copied into the `.prg` output and caused a syntax error in BASIC.

The problem was debugged and fix tested in VICE `x64` on Linux.